fix: replaced magic number of iface with the index for 'eth0'
I used the magic number 2 as the default interface index. This may have come from some forum or blog I read about 2 being the default something interface. This was dangerous code as it was ambiguous and probably unportable. Instead, i replaced that magic value with a call to if_nametoindex() to default for eth0.