new: args struct is now allocated on heap, not stack
I allocated the args structure so that its allocated in the heap, not the stack. I did so that, in the next few commits, I can safely zero out the arg structure since it may contain the password. While I could zero out the arg structure allocated in stack, I didn't feel comfortable touching the stack like that for fear i may corrupt a stack frame.