moved the -c option from the CFLAGS macro to the rule
This is best since the CFLAGS macro will be used in more than one rule. However, these rules do not both need the -c option. Requiring all rules that will use the CFLAGS macro to compile build objects is ridiculous. Therefore, I moved the option to the only specific rule that required it. I'm keeping the CFLAGS macro as general as possible.