compile tests for make 'test' rule
As it were, 'make test' would run run_tests.pl, which, obviously, only executed the tests. If the tests' source files had changed, this change would not be reflected in their respective binaries, unless the sources were manually recompiled. This change compiles the sources before running them to allow the latest sources to be tested