Broadcast across all interfaces.
The way I see it, there are three ways for `wol` to handle the situation where `-i` isn't specified: 1) assume a network interface. This is how it used to behave, but it's error-prone. 2) Make `-i` required, but this changes the behavior and design. 3) Broadcast on all supported interfaces, which is what `wol` will now do.